% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-cardpool.R
\name{build_lineage.netrunneR_git_mirror}
\alias{build_lineage.netrunneR_git_mirror}
\title{Build a git-mirror lineage's processed release}
\usage{
\method{build_lineage}{netrunneR_git_mirror}(lineage, staged_raw, ...)
}
\arguments{
\item{lineage}{A lineage object of class netrunneR_git_mirror.}

\item{staged_raw}{The value returned by fetch_lineage.netrunneR_git_mirror().}

\item{...}{Ignored.}
}
\description{
Dispatches to the cardpool or implementation build routine by lineage
name, since both lineages share the netrunneR_git_mirror class and
fetch method but have distinct schemas and outputs.
}
