% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legality.R
\name{card_pool_sets}
\alias{card_pool_sets}
\title{The card sets making up a card pool}
\usage{
card_pool_sets(card_pool_set, card_pool_id)
}
\arguments{
\item{card_pool_set}{The cardpool \code{card_pool_set} table.}

\item{card_pool_id}{Character. The pool to resolve.}
}
\value{
Character vector of card set ids.
}
\description{
The card sets making up a card pool
}
