% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-helpers.R
\name{check_col_vals_in_set}
\alias{check_col_vals_in_set}
\title{Assert every value of a column falls in an allowed set}
\usage{
check_col_vals_in_set(df, column, allowed)
}
\description{
Assert every value of a column falls in an allowed set
}
