% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-helpers.R
\name{check_rows_distinct}
\alias{check_rows_distinct}
\title{Assert rows are distinct on one or more key columns}
\usage{
check_rows_distinct(df, keys)
}
\description{
Assert rows are distinct on one or more key columns
}
