% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-cobra.R
\name{COBRA_TYPE_ID_COUNT}
\alias{COBRA_TYPE_ID_COUNT}
\title{Number of public tournament-type listing pages Cobra exposes}
\usage{
COBRA_TYPE_ID_COUNT
}
\description{
Number of public tournament-type listing pages Cobra exposes
}
\keyword{internal}
