% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-eval.R
\name{compare_ordered}
\alias{compare_ordered}
\title{Compare two numeric/date vectors under a search operator}
\usage{
compare_ordered(col, value, op)
}
\description{
Compare two numeric/date vectors under a search operator
}
\keyword{internal}
