% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-cobra.R
\name{discover_cobra_recent_index}
\alias{discover_cobra_recent_index}
\title{Scrape the 12 public type pages for currently-listed tournament ids}
\usage{
discover_cobra_recent_index(lineage)
}
\description{
Scrape the 12 public type pages for currently-listed tournament ids
}
\keyword{internal}
