% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download.octgn}
\alias{download.octgn}
\title{Read in an Android: Netrunner OCTGN data file.}
\usage{
download.octgn(quietly = TRUE)
}
\arguments{
\item{quietly}{If TRUE, don't display any messages or a progress bar while downloading.}
}
\description{
This function downloads the latest Android: Netrunner OCTGN data file from http://octgn.gamersjudgment.net/
}

