% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-implementation.R
\name{extract_ice_breaker_traits}
\alias{extract_ice_breaker_traits}
\title{Stub extraction of ice/breaker traits for one card-definition file}
\usage{
extract_ice_breaker_traits(path)
}
\arguments{
\item{path}{Character. Path to a card definition JS file.}
}
\value{
A one-row tibble of ice/breaker trait columns for the card.
}
\description{
Stub extraction of ice/breaker traits for one card-definition file
}
\keyword{internal}
