% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.data.packs}
\alias{get.data.packs}
\title{Get a list of data packs.}
\usage{
get.data.packs()
}
\value{
A data frame listing each data pack with its version number as an ordered factor.
}
\description{
Create a data frame with the data pack as an ordered factor and the version number associated with it. Needs to be updated with each OCTGN version; would be very nice if it scraped db0's GitHub NEWS page and updated itself accordingly.
}

