% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-implementation.R
\name{ice_subroutine_count}
\alias{ice_subroutine_count}
\title{Subroutine count for one ice defcard, or NA}
\usage{
ice_subroutine_count(body)
}
\description{
Subroutine count for one ice defcard, or NA
}
\keyword{internal}
