% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_deck_compare.R
\name{mod_deck_compare_ui}
\alias{mod_deck_compare_ui}
\title{Deck Compare UI}
\usage{
mod_deck_compare_ui(id)
}
\arguments{
\item{id}{Module namespace id.}
}
\value{
A \code{shiny::tagList}.
}
\description{
Deck Compare UI
}
