% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-nrdb.R
\name{NRDB_RULING_ALLOWLIST}
\alias{NRDB_RULING_ALLOWLIST}
\title{Same allowlist, for the ruling table.}
\usage{
NRDB_RULING_ALLOWLIST
}
\description{
Same allowlist, for the ruling table.
}
\keyword{internal}
