% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-clojure.R
\name{read_defcards}
\alias{read_defcards}
\title{Every \verb{(defcard "Title" ...)} form in one file}
\usage{
read_defcards(path)
}
\arguments{
\item{path}{Character. Path to a \code{.clj} card-definition file.}
}
\value{
A list of \code{list(title, body)}, body being the whole form as
text.
}
\description{
Every \verb{(defcard "Title" ...)} form in one file
}
\keyword{internal}
