% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-cardpool.R
\name{read_json_tibble}
\alias{read_json_tibble}
\title{Parse one JSON file into a tibble}
\usage{
read_json_tibble(path)
}
\arguments{
\item{path}{Character. Path to a JSON file.}
}
\value{
A tibble of the parsed JSON content.
}
\description{
Parse one JSON file into a tibble
}
\keyword{internal}
