% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views-ratings.R
\name{run_elo}
\alias{run_elo}
\title{Run an Elo rating pass with fixed identity-elo-v1 parameters}
\usage{
run_elo(games, side_a, side_b, score_a)
}
\description{
Run an Elo rating pass with fixed identity-elo-v1 parameters
}
\keyword{internal}
