% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-syntax.R
\name{search_parse_tokens}
\alias{search_parse_tokens}
\title{Parse a tokenized query into an abstract syntax tree}
\usage{
search_parse_tokens(tokens, fields)
}
\description{
Parse a tokenized query into an abstract syntax tree
}
\keyword{internal}
