% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-nrdb.R
\name{build_nrdb}
\alias{build_nrdb}
\title{Build the nrdb release}
\usage{
build_nrdb(lineage, staged_raw)
}
\arguments{
\item{lineage}{A lineage object of class netrunneR_api_poll named "nrdb".}

\item{staged_raw}{The value returned by fetch_nrdb().}
}
\description{
Builds the nrdb review and ruling tables from the fetched envelopes.
}
\keyword{internal}
