% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_date_filter.R
\name{date_bounds}
\alias{date_bounds}
\title{The span of dates present in the data}
\usage{
date_bounds(dated)
}
\arguments{
\item{dated}{Output of with_parsed_dates().}
}
\value{
A length-2 Date vector, or NULL when nothing parsed.
}
\description{
The span of dates present in the data
}
\keyword{internal}
