% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download.octgn}
\alias{download.octgn}
\title{Read in an Android: Netrunner OCTGN data file.}
\usage{
download.octgn(write = FALSE)
}
\arguments{
\item{write}{If TRUE, write the downloaded file out as a .csv.}
}
\value{
If \code{write = FALSE}, return the downloaded data as a
  data frame. Otherwise, write a CSV file.
}
\description{
This function downloads the latest Android: Netrunner OCTGN .csv data file
from http://octgn.gamersjudgment.net/... or it will when it works.
}

